In computer class

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,11 +8,11 @@ int main(void)
 {
 int i;
 int t1 = 1, t2 = 2;
-long long nextTerm = t1 + t2;
+long int nextTerm = t1 + t2;
 	printf("1, 2, ");
 for (i = 3; i <= 50; ++i)
 {
-	printf("%lld, ", nextTerm);
+	printf("%ld, ", nextTerm);
 	t1 = t2;
 	t2 = nextTerm;
 	nextTerm = t1 + t2;
